Fix 500 error on storage page: cast disk_*_space floats to int
disk_total_space() and disk_free_space() return float, but
formatBytes() expects int. With strict_types=1 enabled, PHP throws
a TypeError instead of silently coercing the value.

+++ src/Controllers/AdminController.php
@@ -1057,8 +1057,8 @@ class AdminController
         $basePath = \SDS\Core\App::basePath();
 
         // Total Drive (/ filesystem)
-        $totalBytes = @disk_total_space('/') ?: 0;
-        $freeBytes  = @disk_free_space('/') ?: 0;
+        $totalBytes = (int) (@disk_total_space('/') ?: 0);
+        $freeBytes  = (int) (@disk_free_space('/') ?: 0);
         $usedBytes  = $totalBytes - $freeBytes;
 
         $categories = [];
